currency e2e
