+++ test/pages/modules/consent_mgt_gdpr.html
@@ -1,6 +1,6 @@
 <html>
   <head>
-    <script async src="../../build/dist/prebid.js"></script>
+    <script async src="../../../build/dist/prebid.js"></script>
     <script async src="https://www.googletagservices.com/tag/js/gpt.js"></script>
     <script>
       (function(window, document) {
